+++ src/main/java/seedu/algobase/logic/commands/Command.java
@@ -2,6 +2,20 @@ package seedu.algobase.logic.commands;
 
 import seedu.algobase.logic.CommandHistory;
 import seedu.algobase.logic.commands.exceptions.CommandException;
+import seedu.algobase.logic.commands.problem.AddCommand;
+import seedu.algobase.logic.commands.problem.DeleteCommand;
+import seedu.algobase.logic.commands.problem.EditCommand;
+import seedu.algobase.logic.commands.problem.FindCommand;
+import seedu.algobase.logic.commands.problem.ListCommand;
+import seedu.algobase.logic.commands.problem.SortCommand;
+import seedu.algobase.logic.commands.storage.ExportCommand;
+import seedu.algobase.logic.commands.storage.ImportCommand;
+import seedu.algobase.logic.commands.task.AddTaskCommand;
+import seedu.algobase.logic.commands.task.DeleteTaskCommand;
+import seedu.algobase.logic.commands.task.DoneTaskCommand;
+import seedu.algobase.logic.commands.task.EditTaskCommand;
+import seedu.algobase.logic.commands.task.MoveTaskCommand;
+import seedu.algobase.logic.commands.task.UndoneTaskCommand;
 import seedu.algobase.model.Model;
 
 /**
