+++ src/main/java/seedu/algobase/logic/parser/DeleteTagCommandParser.java
@@ -3,8 +3,8 @@ package seedu.algobase.logic.parser;
 import static seedu.algobase.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.algobase.commons.core.index.Index;
-import seedu.algobase.logic.commands.DeleteCommand;
 import seedu.algobase.logic.commands.DeleteTagCommand;
+import seedu.algobase.logic.commands.problem.DeleteCommand;
 import seedu.algobase.logic.parser.exceptions.ParseException;
 
 /**
