+++ src/main/java/seedu/algobase/logic/parser/ParserUtil.java
@@ -21,8 +21,8 @@ import seedu.algobase.commons.core.index.Index;
 import seedu.algobase.commons.util.FileUtil.Format;
 import seedu.algobase.commons.util.StringUtil;
 import seedu.algobase.logic.commands.OpenTabCommand;
-import seedu.algobase.logic.commands.SortCommand;
 import seedu.algobase.logic.commands.SwitchTabCommand;
+import seedu.algobase.logic.commands.problem.SortCommand;
 import seedu.algobase.logic.parser.exceptions.ParseException;
 import seedu.algobase.model.ModelType;
 import seedu.algobase.model.gui.TabType;
