+++ src/main/java/seedu/algobase/logic/parser/task/DoneTaskCommandParser.java
@@ -1,30 +1,34 @@
-package seedu.algobase.logic.parser;
+package seedu.algobase.logic.parser.task;
 
 import static seedu.algobase.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.algobase.logic.parser.CliSyntax.PREFIX_PLAN;
 import static seedu.algobase.logic.parser.CliSyntax.PREFIX_TASK;
 
 import seedu.algobase.commons.core.index.Index;
-import seedu.algobase.logic.commands.UndoneTaskCommand;
+import seedu.algobase.logic.commands.task.DoneTaskCommand;
+import seedu.algobase.logic.parser.ArgumentMultimap;
+import seedu.algobase.logic.parser.ArgumentTokenizer;
+import seedu.algobase.logic.parser.Parser;
+import seedu.algobase.logic.parser.ParserUtil;
 import seedu.algobase.logic.parser.exceptions.ParseException;
 
 /**
  * Parses input arguments and creates a new DeleteTaskCommand object
  */
-public class UndoneTaskCommandParser implements Parser<UndoneTaskCommand> {
+public class DoneTaskCommandParser implements Parser<DoneTaskCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the DoneTaskCommand
      * and returns a DoneTaskCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public UndoneTaskCommand parse(String args) throws ParseException {
+    public DoneTaskCommand parse(String args) throws ParseException {
         ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args, PREFIX_PLAN, PREFIX_TASK);
 
         if (!ParserUtil.arePrefixesPresent(argMultimap, PREFIX_PLAN)
             || !ParserUtil.arePrefixesPresent(argMultimap, PREFIX_TASK)
             || !argMultimap.getPreamble().isEmpty()) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, UndoneTaskCommand.MESSAGE_USAGE));
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DoneTaskCommand.MESSAGE_USAGE));
         }
 
         Index planIndex;
@@ -32,7 +36,7 @@ public class UndoneTaskCommandParser implements Parser<UndoneTaskCommand> {
             planIndex = ParserUtil.parseIndex(argMultimap.getValue(PREFIX_PLAN).get());
         } catch (ParseException pe) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                UndoneTaskCommand.MESSAGE_USAGE), pe);
+                DoneTaskCommand.MESSAGE_USAGE), pe);
         }
 
         Index taskIndex;
@@ -40,13 +44,13 @@ public class UndoneTaskCommandParser implements Parser<UndoneTaskCommand> {
             taskIndex = ParserUtil.parseIndex(argMultimap.getValue(PREFIX_TASK).get());
         } catch (ParseException pe) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                UndoneTaskCommand.MESSAGE_USAGE), pe);
+                DoneTaskCommand.MESSAGE_USAGE), pe);
         }
 
-        UndoneTaskCommand.UndoneTaskDescriptor doneTaskDescriptor =
-            new UndoneTaskCommand.UndoneTaskDescriptor(planIndex, taskIndex);
+        DoneTaskCommand.DoneTaskDescriptor doneTaskDescriptor =
+            new DoneTaskCommand.DoneTaskDescriptor(planIndex, taskIndex);
 
-        return new UndoneTaskCommand(doneTaskDescriptor);
+        return new DoneTaskCommand(doneTaskDescriptor);
     }
 
 }
