+++ src/test/java/seedu/algobase/logic/commands/AddCommandTest.java
@@ -11,6 +11,7 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import seedu.algobase.logic.CommandHistory;
+import seedu.algobase.logic.commands.problem.AddCommand;
 import seedu.algobase.model.AlgoBase;
 import seedu.algobase.model.ReadOnlyAlgoBase;
 import seedu.algobase.model.problem.Problem;
