+++ src/test/java/seedu/algobase/logic/commands/HelpCommandTest.java
@@ -11,6 +11,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import seedu.algobase.logic.CommandHistory;
+import seedu.algobase.logic.commands.problem.AddCommand;
 import seedu.algobase.model.Model;
 import seedu.algobase.model.ModelManager;
 import seedu.algobase.model.UserPrefs;
